data/requirements.txt
pandas
boto3
pyarrow
python-dotenv
fsspec
s3fs